解决合并代码冲突

+++ xwas-web/src/main/java/cn/com/xinxin/sass/web/rest/SassResourceRestController.java
@@ -1,31 +1,62 @@
 package cn.com.xinxin.sass.web.rest;
 
+/*
+ *
+ * Copyright 2020 www.xinxindigits.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ * and associated documentation files (the "Software"),to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice
+ * shall be included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * Redistribution and selling copies of the software are prohibited, only if the authorization from xinxin digits
+ * was obtained.Neither the name of the xinxin digits; nor the names of its contributors may be used to
+ * endorse or promote products derived from this software without specific prior written permission.
+ *
+ */
+
+import cn.com.xinxin.sass.auth.enums.AuthEnums;
 import cn.com.xinxin.sass.auth.model.SassUserInfo;
 import cn.com.xinxin.sass.auth.web.AclController;
-import cn.com.xinxin.sass.biz.service.ResourceService;
-import cn.com.xinxin.sass.biz.service.RoleResourceService;
-import cn.com.xinxin.sass.biz.service.RoleService;
-import cn.com.xinxin.sass.biz.service.UserService;
+import cn.com.xinxin.sass.biz.log.SysLog;
+import cn.com.xinxin.sass.biz.service.*;
 import cn.com.xinxin.sass.common.enums.ResourceTypeEnums;
 import cn.com.xinxin.sass.common.enums.SassBizResultCodeEnum;
 import cn.com.xinxin.sass.common.model.PageResultVO;
+import cn.com.xinxin.sass.repository.model.AuthsDO;
 import cn.com.xinxin.sass.repository.model.ResourceDO;
 import cn.com.xinxin.sass.repository.model.RoleResourceDO;
 import cn.com.xinxin.sass.web.convert.SassFormConvert;
 import cn.com.xinxin.sass.web.form.ResourceForm;
 import cn.com.xinxin.sass.web.form.ResourceQueryForm;
+import cn.com.xinxin.sass.web.utils.RegexUtils;
 import cn.com.xinxin.sass.web.utils.TreeResultUtil;
+import cn.com.xinxin.sass.web.vo.GrantsVO;
 import cn.com.xinxin.sass.web.vo.MenuTreeVO;
 import cn.com.xinxin.sass.web.vo.ResourceVO;
 import com.google.common.collect.Lists;
 import com.xinxinfinance.commons.exception.BusinessException;
+import com.xinxinfinance.commons.util.BaseConvert;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.annotations.Param;
+import org.apache.shiro.authz.annotation.Logical;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -57,10 +88,14 @@ public class SassResourceRestController extends AclController {
     @Autowired
     private RoleResourceService roleResourceService;
 
+    @Autowired
+    private AuthsService authsService;
+
 
     @RequestMapping(value = "/list",method = RequestMethod.POST)
     @ResponseBody
-    @RequiresPermissions("/resource/list")
+    @RequiresPermissions(value = {"SASS_RESOURCE_MENU_FUNCTION_MNG",
+            "SASS_RESOURCE_MENU_QUERY_LIST"},logical= Logical.OR)
     public Object ListResource(HttpServletRequest request, @RequestBody ResourceQueryForm resourceQueryForm){
 
         log.info("ResourceController.list,resourceQueryForm={}", resourceQueryForm);
@@ -85,6 +120,38 @@ public class SassResourceRestController extends AclController {
 
         return result;
     }
+
+
+    @RequestMapping(value = "/grants/list",method = RequestMethod.GET)
+    @ResponseBody
+    public Object listAllGrantsValues(HttpServletRequest request){
+
+        log.info("ResourceController.listAllGrantsValues");
+
+//        List<AuthEnums> authEnumsList = AuthEnums.toLists();
+//        List<GrantsVO> allGrants = Lists.newArrayList();
+//        authEnumsList.stream().forEach(authEnums -> {
+//            GrantsVO grantsVO = new GrantsVO();
+//            grantsVO.setCode(authEnums.getCode());
+//            grantsVO.setName(authEnums.getName());
+//            grantsVO.setDesc(authEnums.getDesc());
+//            allGrants.add(grantsVO);
+//
+//        });
+
+        List<AuthsDO> authsDOList = this.authsService.selectAllAuths();
+        List<GrantsVO> allGrants = Lists.newArrayList();
+        authsDOList.stream().forEach(authsDO -> {
+            GrantsVO grantsVO = new GrantsVO();
+            grantsVO.setCode(authsDO.getAuthority());
+            grantsVO.setName(authsDO.getName());
+            grantsVO.setDesc(authsDO.getName());
+            allGrants.add(grantsVO);
+        });
+
+        return allGrants;
+    }
+
 
 
 
@@ -126,7 +193,6 @@ public class SassResourceRestController extends AclController {
      */
     @RequestMapping(value = "/menu/tree",method = RequestMethod.GET)
     @ResponseBody
-    //@RequiresPermissions("resource/menu/tree")
     public Object treeMenuResource(HttpServletRequest request){
 
         log.info("ResourceController.treeMenuResource,roleCode={}");
@@ -156,6 +222,8 @@ public class SassResourceRestController extends AclController {
                     menuTreeVO.setCode(resourceVO.getCode());
                     menuTreeVO.setUrl(resourceVO.getUrl());
                     menuTreeVO.setAuthority(resourceVO.getAuthority());
+                    menuTreeVO.setExtension(resourceVO.getExtension());
+                    menuTreeVO.setChecked(false);
                     menuTreeVO.setOrder(0);
                     resourceTreeVOList.add(menuTreeVO);
                 }
@@ -177,7 +245,6 @@ public class SassResourceRestController extends AclController {
      */
     @RequestMapping(value = "/tree",method = RequestMethod.GET)
     @ResponseBody
-    @RequiresPermissions("/resource/tree")
     public Object treeResource(HttpServletRequest request,
                                @Param("roleCode") String roleCode){
 
@@ -213,6 +280,8 @@ public class SassResourceRestController extends AclController {
                     menuTreeVO.setCode(resourceVO.getCode());
                     menuTreeVO.setUrl(resourceVO.getUrl());
                     menuTreeVO.setAuthority(resourceVO.getAuthority());
+                    menuTreeVO.setExtension(resourceVO.getExtension());
+                    menuTreeVO.setChecked(false);
                     menuTreeVO.setOrder(0);
                     resourceTreeVOList.add(menuTreeVO);
                 }
@@ -227,9 +296,17 @@ public class SassResourceRestController extends AclController {
 
     @RequestMapping(value = "/create",method = RequestMethod.POST)
     @ResponseBody
-    @RequiresPermissions("/resource/create")
+    @RequiresPermissions(value = {"SASS_RESOURCE_MENU_FUNCTION_MNG",
+            "SASS_RESOURCE_MENU_FUNCTION_ADD"},logical= Logical.OR)
+    @SysLog("创建资源权限")
     public Object createResource(HttpServletRequest request,
                                    @RequestBody ResourceForm resourceForm){
+
+        if(!RegexUtils.isDataCode(resourceForm.getCode())){
+            // 如果匹配不是useraccount格式
+            throw new BusinessException(SassBizResultCodeEnum.ILLEGAL_PARAMETER,
+                    "编码不能包含特殊字符或者长度超过16","编码不能包含特殊字符或者长度超过16");
+        }
 
         SassUserInfo sassUserInfo = this.getSassUser(request);
         String userAccount = sassUserInfo.getAccount();
@@ -238,18 +315,41 @@ public class SassResourceRestController extends AclController {
         ResourceDO resourceDO = SassFormConvert.convertResourceForm2ResourceDO(resourceForm);
         resourceDO.setGmtCreator(userAccount);
         resourceDO.setGmtUpdater(userAccount);
-        ResourceDO resultDO = resourceService.createResource(resourceDO);
 
-        if(null != resultDO){
-            return SassBizResultCodeEnum.SUCCESS.getAlertMessage();
-        }else{
-            throw new BusinessException(SassBizResultCodeEnum.FAIL,"创建资源出错","清检查参数是否正确");
+        // FIXME: 先默认设置为xinxin租户
+        resourceDO.setTenantId("xinxin");
+
+
+        try {
+            int result = resourceService.createResource(resourceDO);
+
+            if(result==0){
+                throw new BusinessException(SassBizResultCodeEnum.FAIL,"创建资源出错","清检查参数是否正确");
+            }
+
+        }catch (DuplicateKeyException dex){
+
+            throw new BusinessException(SassBizResultCodeEnum.ILLEGAL_PARAMETER, "编码不能重复","编码不能重复");
+
+        }catch (Exception ex){
+
+            throw new BusinessException(SassBizResultCodeEnum.FAIL, "处理异常，请稍后重试","处理异常，请稍后重试");
+
         }
+
+        ResourceDO createdResourceDO = this.resourceService.findByResourceCode(resourceForm.getCode());
+
+        ResourceVO createResVO = BaseConvert.convert(createdResourceDO, ResourceVO.class);
+
+        return createResVO;
+
     }
 
+    @SysLog("更新资源权限")
     @RequestMapping(value = "/update",method = RequestMethod.POST)
     @ResponseBody
-    @RequiresPermissions("/resource/update")
+    @RequiresPermissions(value = {"SASS_RESOURCE_MENU_FUNCTION_MNG",
+            "SASS_RESOURCE_MENU_FUNCTION_UPDATE"},logical= Logical.OR)
     public Object updateResource(HttpServletRequest request,
                              @RequestBody ResourceForm resourceForm){
 
@@ -267,10 +367,11 @@ public class SassResourceRestController extends AclController {
 
     }
 
-
+    @SysLog("删除资源权限")
     @RequestMapping(value = "/delete",method = RequestMethod.DELETE)
     @ResponseBody
-    @RequiresPermissions("/resource/delete")
+    @RequiresPermissions(value = {"SASS_RESOURCE_MENU_FUNCTION_MNG",
+            "SASS_RESOURCE_MENU_FUNCTION_REMOVE"},logical= Logical.OR)
     public Object deleteResource(HttpServletRequest request,
                                  @RequestParam String id){
 
@@ -288,6 +389,14 @@ public class SassResourceRestController extends AclController {
             // 如果某个资源关联角色，则不能删除该权限资源
             throw new BusinessException(SassBizResultCodeEnum.FAIL,
                     "该资源权限已经使用，删除需要移除关联角色","该资源权限已经使用，删除需要移除关联角色");
+        }
+
+        List<ResourceDO> subResourceDOList =
+                this.resourceService.findChildrenByParentIds(Lists.newArrayList(Long.valueOf(id)));
+
+        if(CollectionUtils.isNotEmpty(subResourceDOList)){
+            throw new BusinessException(SassBizResultCodeEnum.FAIL,
+                    "该资源权限已经使用，并且关联其他资源类型无法删除","该资源权限已经使用，并且关联其他资源类型无法删除");
         }
 
         boolean result = resourceService.deleteById(Long.valueOf(id));
